cleanup the codes
