+++ lmdeploy/archs.py
@@ -103,13 +103,19 @@ def autoget_backend_config(
     return config
 
 
-def check_vl_llm(arch: str, config: dict) -> bool:
+def check_vl_llm(config: dict) -> bool:
     """check if the model is a vl model from model config."""
     if 'auto_map' in config:
         for _, v in config['auto_map'].items():
             if 'InternLMXComposer2ForCausalLM' in v:
                 return True
 
+    if 'language_config' in config and \
+       'vision_config' in config and \
+       config.get('language_config').get('architectures', [None])[0] == 'DeepseekV2ForCausalLM':
+        return True
+
+    arch = config['architectures'][0]
     supported_archs = set([
         'LlavaLlamaForCausalLM', 'LlavaMistralForCausalLM', 'CogVLMForCausalLM', 'InternLMXComposer2ForCausalLM',
         'InternVLChatModel', 'MiniGeminiLlamaForCausalLM', 'MGMLlamaForCausalLM', 'MiniCPMV',
@@ -121,8 +127,6 @@ def check_vl_llm(arch: str, config: dict) -> bool:
     elif arch == 'MultiModalityCausalLM' and 'language_config' in config:
         return True
     elif arch in ['ChatGLMModel', 'ChatGLMForConditionalGeneration'] and 'vision_config' in config:
-        return True
-    elif arch == 'DeepseekV2ForCausalLM' and 'vision_config' in config:
         return True
     elif arch in supported_archs:
         return True
@@ -136,8 +140,8 @@ def get_task(model_path: str):
     if os.path.exists(os.path.join(model_path, 'triton_models', 'weights')):
         # workspace model
         return 'llm', AsyncEngine
-    arch, config = get_model_arch(model_path)
-    if check_vl_llm(arch, config.to_dict()):
+    _, config = get_model_arch(model_path)
+    if check_vl_llm(config.to_dict()):
         from lmdeploy.serve.vl_async_engine import VLAsyncEngine
         return 'vlm', VLAsyncEngine
