+++ lmdeploy/model.py
@@ -1475,6 +1475,8 @@ class DeepseekVL2(BaseChatTemplate):
         return super().get_prompt(prompt, sequence_start)[:-1]
 
     def messages2prompt(self, messages, sequence_start=True, **kwargs):
+        if isinstance(messages, str):
+            return self.get_prompt(messages, sequence_start)
         return super().messages2prompt(messages, sequence_start, **kwargs)[:-1]
 
     @classmethod
